Make RenderScene the work dispatching class

+++ src/swrenderer/drawers/r_thread.cpp
@@ -282,5 +282,5 @@ DrawerCommandQueue::DrawerCommandQueue(swrenderer::RenderThread *renderthread) :
 
 void *DrawerCommandQueue::AllocMemory(size_t size)
 {
-	return renderthread->FrameMemory->AllocMemory<uint8_t>(size);
+	return renderthread->FrameMemory->AllocMemory<uint8_t>((int)size);
 }
